requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-login==0.6.3
werkzeug==2.3.7
bcrypt==4.0.1
flask-talisman==1.0.0
flask-limiter==3.5.1
